Translation Updates (Bosnian, Dutch, British English), along with fixes in German (#15)
The en-UK file is a bit more different than the original, using more text in some areas to make it clearer what it means (along with commas in certain places). The only real significant difference is changing "Licenses" to "Licences".

I fixed a lot of parts in German, and filled in the blank spaces the other contributor left behind.

Bosnian and Dutch are the absolute newest additions. I would've also added Serbian and Croatian, but they're not necessary unless requested (as they're (along with Bosnian) all pretty much one and the same language).